+++ src/app/_components/template/explore/Template.style.tsx
@@ -1,0 +1,8 @@
+import { HomeLayoutStyle } from "@/app/_styles/common.style";
+import styled from "styled-components";
+
+export const Template = {
+  Main: styled(HomeLayoutStyle.Main)`
+    width: auto;
+  `,
+};
