+++ src/app/_components/template/home/HomeTemplate.tsx
@@ -1,7 +1,7 @@
 import { HomeLayoutStyle } from "@/app/_styles/common.style";
 import React from "react";
-import StoryBar from "../organism/storyBar/StoryBar";
-import FeedList from "../organism/feedlist/FeedList";
+import StoryBar from "../../organism/storyBar/StoryBar";
+import FeedList from "../../organism/feedlist/FeedList";
 
 export default function HomeTemplate() {
   return (
